FitHiChIP loop calls from human brain tissue

FitHiChIP loop calls that overlap SNPs derived from analysis of H3K27ac HiChIP data. Each row represents an individual peak identified from the feature binarization analysis (see methods).

CORCES_2020.HiChIP_FitHiChIP_loop_calls

Format

An object of class data.table (inherits from data.frame) with 11542 rows and 11 columns.

Source

https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1

Details

Data originally from Corces et al. (bioRxiv), as of May 2020. Specifically: STable10_Coacessibility_Peak_loop_connection, HiChIP FitHiChIP Loop Calls sheet.

Column dictionary

hg38_Chromosome_Anchor1

The hg38 chromosome of the first loop Anchor.

hg38_Start_Anchor1

The hg38 start position of the first loop Anchor.

hg38_Stop_Anchor1

The hg38 stop position of the first loop Anchor.

Width_Anchor1

The width of the first loop Anchor.

hg38_Chromosome_Anchor2

The hg38 chromosome of the second loop Anchor.

hg38_Start_Anchor2

The hg38 start position of the second loop Anchor.

hg38_Stop_Anchor2

The hg38 stop position of the second loop Anchor.

Width_Anchor2

The width of the second loop Anchor.

Score

The -log10(q-value) of the loop call from FitHiChIP.

Anchor1_hasSNP

A boolean variable determining whether the first anchor overlaps a SNP from our AD/PD GWAS analyses.

Anchor2_hasSNP

A boolean variable determining whether the second anchor overlaps a SNP from our AD/PD GWAS analyses.

See also

Examples

if (FALSE) {
dat <- readxl::read_excel("~/Desktop/Fine_Mapping/echolocatoR/annotations/Coceres_2020/STable10_Coacessibility_Peak_loop_connection.xlsx", skip = 19, sheet=1)
CORCES_2020.HiChIP_FitHiChIP_loop_calls <- data.table::data.table(dat)
usethis::use_data(CORCES_2020.HiChIP_FitHiChIP_loop_calls)
}