Cicero_coaccessibility from Alzheimer's disease brain tissue

Cicero coaccessibility analysis for peaks that overlap SNPs derived from analysis of scATAC-seq data. Each row represents an individual peak identified from the feature binarization analysis (see methods).

CORCES_2020.cicero_coaccessibility

Format

An object of class data.table (inherits from data.frame) with 9795 rows and 14 columns.

Source

https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1

Details

Data originally from Corces et al. (bioRxiv), as of May 2020. Specifically: STable10_Coacessibility_Peak_loop_connection, Cicero Coaccessibility sheet. Peak_ID_Peak1 - A unique number that identifies the peak across supplementary tables.

Column dictionary:

hg38_Chromosome_Peak1

The hg38 chromosome of the first loop Peak.

hg38_Start_Peak1

The hg38 start position of the first loop Peak.

hg38_Stop_Peak1

The hg38 stop position of the first loop Peak.

Width_Peak1

The width of the first loop Peak.

Peak_ID_Peak2

A unique number that identifies the peak across supplementary tables.

hg38_Chromosome_Peak2

The hg38 chromosome of the second loop Peak.

hg38_Start_Peak2

The hg38 start position of the second loop Peak.

hg38_Stop_Peak2

The hg38 stop position of the second loop Peak.

Width_Peak2

The width of the second loop Peak.

Coaccessibility

The coaccessibility correlation for the given peak pair.

Peak1_hasSNP

A boolean variable determining whether the first peak overlaps a SNP from our AD/PD GWAS analyses.

Peak2_hasSNP

A boolean variable determining whether the second peak overlaps a SNP from our AD/PD GWAS analyses.

See also

Examples

if (FALSE) {
dat <- readxl::read_excel("~/Desktop/Fine_Mapping/echolocatoR/annotations/Coceres_2020/STable10_Coacessibility_Peak_loop_connection.xlsx", skip = 21, sheet=2)
CORCES_2020.cicero_coaccessibility <- data.table::data.table(dat)
usethis::use_data(CORCES_2020.cicero_coaccessibility)
}