Multi-track plot of fine-mapping results

Plots fine-mapping results from echolocatoR. Create separate tracks for the original results (-log10(P)) and each fine-mapping method (Posterior Probability). Uses ggbio.

GGBIO.plot(
  finemap_dat,
  locus_dir,
  LD_matrix = NULL,
  color_r2 = T,
  method_list = c("ABF", "FINEMAP", "SUSIE", "POLYFUN_SUSIE"),
  label_snps = T,
  dot_summary = F,
  QTL_prefixes = NULL,
  mean.PP = T,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  sig_cutoff = 5e-08,
  XGR_libnames = c("ENCODE_TFBS_ClusteredV3_CellTypes",
    "ENCODE_DNaseI_ClusteredV3_CellTypes", "Broad_Histone"),
  n_top_xgr = 5,
  Roadmap = F,
  Roadmap_query = NULL,
  n_top_roadmap = 7,
  annot_overlap_threshold = 5,
  Nott_epigenome = F,
  Nott_regulatory_rects = T,
  Nott_show_placseq = T,
  Nott_binwidth = 200,
  Nott_bigwig_dir = NULL,
  save_plot = T,
  show_plot = T,
  max_transcripts = 1,
  plot.window = NULL,
  dpi = 300,
  height = 12,
  width = 10,
  verbose = T
)

Arguments

color_r2
If an LD_matrix is supplied, color each SNP by its LD with the lead SNP from the GWAS ()

QTL_prefixes
Prefixes to the columns that contain QTL data. For exmaple, P-values for the Fairfax_2014 QTL study would be stored in the column Fairfax_2014.P.

PP_threshold
The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set. For example, PP_threshold=.95 means that all Credible Set SNPs will be 95% Credible Set SNPs.

consensus_threshold
The minimum number of fine-mapping tools that include a SNP in their 95% Credible Sets to consider that it a "Consensus SNP" (default=2).

max_transcripts
Max number of transcripts per gene to show.

plot.window
Zoom into the center of the locus when plotting (without editing the fine-mapping results file). You can also pass a list of window sizes (e.g. c(50000,100000,500000)) to automatically generate multiple views of each locus.

verbose
Whether echolocatoR should be verbose or silent.

Source

http://bioconductor.org/packages/release/bioc/vignettes/ggbio/inst/doc/ggbio.pdf

See also

Examples