Gather all IMPACT annotations that overlap with your query

Iterates over each unique locus. WARNING! These files are quite large and you need to make sure you have ample memory and storage on your computer for best results.

IMPACT.iterate_get_annotations(
  merged_DT,
  IMPACT_score_thresh = 0.1,
  baseURL = "https://github.com/immunogenomics/IMPACT/raw/master/IMPACT707/Annotations",
  all_snps_in_range = T,
  top_annotations = F,
  force_one_annot_per_locus = F,
  snp.filter = "!is.na(SNP)",
  nThread = 4,
  verbose = T
)

Details

Unfortunately, you have to download the entire chromosome file at once, because they aren't Tabix indexed. To minimize the memory load, this function only keeps the portion of the IMPACT file that overlaps with the coordinates in subset_DT.

See also

Examples

if (FALSE) {
data("merged_DT")
ANNOT_MELT <- IMPACT.iterate_get_annotations(merged_DT=merged_DT)
}