Compute LD from 1000 Genomes

Downloads a subset vcf of the 1KG database that matches your locus coordinates. Then uses plink to calculate LD on the fly.

LD.1KG(
  locus_dir,
  subset_DT,
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  vcf_folder = NULL,
  remote_LD = T,
  LD_block = F,
  LD_block_size = 0.7,
  remove_correlates = F,
  remove_tmps = T,
  fillNA = 0,
  download_method = "wget",
  nThread = 4,
  conda_env = "echoR",
  verbose = T
)

Arguments

LD_reference
Which linkage disequilibrium reference panel do you want to use. Options include:

"UKB"

A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See Wiessbrod et al. (2019) for more details.

"1KGphase1"

Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.

"1KGphase3"

Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.

"<path>/*.vcf" or "<path>/*.vcf.gz"

Alternatively, users can provide their own custom panel by supplying a list of .vcf file path (one per locus) which echolocatoR will use to compute LD (using plink).

superpopulation
Subset your LD reference panel by superopulation. Setting the superpopulation is not currently possible when LD_reference="UKB". 1KGphase1 options include:

"AFR"

African [descent]

"AMR"

Ad-mixed American

"EAS"

East Asian

"EUR"

European

"SAS"

South Asian

remote_LD
When acquiring LD matrixes, the default is to delete the full vcf or npz files after echolocatoR has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set remote_LD=T.

LD_block
Calculate LD blocks with plink and only include the block to which the lead SNP belongs.

LD_block_size
Adjust the granularity of block sizes when LD_block=T.

remove_correlates
A named list, where the names are the RSIDs of SNPs whose LD correlates you wish to remove, and the value is the absolute r2 threshold you wish to filter at for each RSID respectively (e.g. remove_correlates = c("rs76904798"=.2, "rs10000737"=.8)). This will also remove the SNPs in remove_correlates themselves.

remove_tmps
Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.

fillNA
When pairwise LD (r) between two SNPs is NA, replace with 0.

conda_env
The name of a conda environment to use.

verbose
Whether echolocatoR should be verbose or silent.

Details

This approach is taken, because other API query tools have limitations with the window size being queried. This approach does not have this limitations, allowing you to fine-map loci more completely.

See also

Examples

if (FALSE) {
data("BST1"); data("locus_dir");
BST1 <- limit_SNPs(max_snps = 500, subset_DT = BST1)
LD_matrix <- LD.1KG(locus_dir=file.path("~/Desktop",locus_dir), subset_DT=BST1, LD_reference="1KGphase1")

## Kunkle et al 2019
locus_dir <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ACE"

}