Download vcf subset from 1000 Genomes

Download vcf subset from 1000 Genomes

LD.1KG_download_vcf(
  subset_DT,
  LD_reference = "1KGphase1",
  vcf_folder = NULL,
  locus_dir,
  locus = NULL,
  download_reference = T,
  whole_vcf = F,
  download_method = "wget",
  force_new_vcf = F,
  nThread = 4,
  query_by_regions = F,
  verbose = T
)

Arguments

LD_reference
Which linkage disequilibrium reference panel do you want to use. Options include:

"UKB"

A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See Wiessbrod et al. (2019) for more details.

"1KGphase1"

Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.

"1KGphase3"

Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.

download_reference
When acquiring LD matrixes, the default is to delete the full vcf or npz files after echolocatoR has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set download_reference=T.

query_by_regions
You can make queries with tabix in two different ways:

query_by_regions=F (default)

Return a vcf with all positions between the min/max in subset_DT Takes up more storage but is MUCH faster

query_by_regions=T

Return a vcf with only the exact positions present in subset_DT. Takes up less storage but is MUCH slower

verbose
Whether echolocatoR should be verbose or silent.

See also

Examples

if (FALSE) {
data("BST1");
subset_DT <- BST1
subset_vcf.popDat <- LD.1KG_download_vcf(subset_DT=BST1, LD_reference="1KGphase1", locus_dir=file.path("~/Desktop",locus_dir))
}