Create LD matrix from plink output.

Depending on which parameters you give plink when calculating LD, you get different file outputs. When it produces an LD table, use this function to create a proper LD matrix.

LD.read_ld_table(ld.path, snp.subset = F, fillNA = 0, verbose = T)

See also