Run motifbreakR

motifbreakR is a package to predict how much a SNP will disrupt a transcription factor binding motif (if it falls witihn one).

MOTIFBREAKR(
  rsid_list,
  save_rds = T,
  dataset_dir = "./results",
  pwmList = NULL,
  organism = NULL,
  threshold = 0.85,
  show.neutral = F,
  method = "default",
  verbose = T,
  calculate_all_pval = T,
  force_new = F
)

Source

Publication: https://pubmed.ncbi.nlm.nih.gov/26272984/ GitHub: https://github.com/Simon-Coetzee/MotifBreakR Vignette: http://simon-coetzee.github.io/motifBreakR

Value

motifbreakr results

See also

Examples

if (FALSE) {
# data("merged_DT")
# rsid_list <- unique(subset(merged_DT, Locus %in% c("LRRK2","MED12L","DYRK1A","FCGR2A") & (Consensus_SNP | leadSNP))$SNP)
# rsid_list <- unique(subset(merged_DT, Consensus_SNP | leadSNP)$SNP)
rsid_list <- c("rs11175620","rs7294619","rs74324737")
mb.results <- MOTIFBREAKR(rsid_list=rsid_list, calculate_all_pval=T, force_new = T)
}