Summarise motifbreakR + echolocatoR results

For each SNP we have at least one allele achieving a p-value below 1e-4 threshold that we required. The seqMatch column shows what the reference genome sequence is at that location, with the variant position appearing in an uppercase letter. pctRef and pctAlt display the the score for the motif in the sequence as a percentage of the best score that motif could achieve on an ideal sequence. In other words (scoreVariant−minscorePWM)/(maxscorePWM−minscorePWM). We can also see the absolute scores for our method in scoreRef and scoreAlt and thier respective p-values.

MOTIFBREAKR.filter(
  merged_DT,
  mb.results,
  pct_threshold = NULL,
  pvalue_threshold = 1e-04,
  qvalue_threshold = 0.05,
  effect_strengths = c("strong"),
  snp_filter = "Consensus_SNP==T",
  top_TF_hits = F,
  no_no_loci = NULL,
  verbose = T
)

Examples

if (FALSE) {
data("merged_DT")
microglia_TF <- read.csv("~/Desktop/Fine_Mapping/resources/microglia_TF.csv")

root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/motifbreakR"
# mb.results <- readRDS(file.path(root, "motifbreakR_results.rds"))
mb.results <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/motifbreakR/motifbreakR_results.rds")

mb.lrrk2 <- readRDS(file.path(root, "motifbreakR_results.p_values_LRRK2.rds"))
mb.encode <- readRDS(file.path(root, "motifbreakR_results.encode.lrrk2.rds"))
mb.DYRK1A_FCGR2A <- readRDS(file.path(root,"mb.results_p.DYRK1A_FCGR2A.RDS"))
mb.MED12L<- readRDS(file.path(root,"MED12L.pvalues.RDS"))

mb.sub <- subset(mb.results, SNP_id %in% subset(merged_DT, Locus=="DNAH17" & (leadSNP | Consensus_SNP))$SNP)
mb.DNAH17 <- motifbreakR::calculatePvalue(results=mb.sub); saveRDS(mb.DNAH17, file.path(root,"DNAH17.pvalues.RDS"))
mb.DNAH17 <- readRDS(file.path(root,"DNAH17.pvalues.RDS"))

mb.MBNL2 <- readRDS(file.path(root, "motifbreakR_results.p_values_MBNL2.rds"))

MOTIFBREAKR.filter(merged_DT, mb.lrrk2, pct_threshold=NULL, effect_strengths=NULL)
}