Plot brain cell-specific epigenomic data

Brain cell-specific epigenomic data from Nott et al. (2019).

NOTT_2019.epigenomic_histograms(
  finemap_dat,
  locus_dir,
  show_plot = T,
  save_plot = T,
  full_data = T,
  return_assay_track = F,
  binwidth = 200,
  density_adjust = 0.2,
  plot.zoom = "1x",
  strip.text.y.angle = 90,
  xtext = T,
  geom = "density",
  plot_formula = "Cell_type ~.",
  fill_var = "Assay",
  bigwig_dir = NULL,
  genomic_units = "Mb",
  as_ggplot = T,
  nThread = 4,
  save_annot = F,
  verbose = T
)

Source

Nott et al. (2019) https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf

Arguments

plot.zoom

Zoom into the center of the locus when plotting (without editing the fine-mapping results file). You can provide either:

The size of your plot window in terms of basepairs (e.g. plot.zoom=50000 for a 50kb window).

How much you want to zoom in (e.g. plot.zoom="1x" for the full locus, plot.zoom="2x" for 2x zoom into the center of the locus, etc.).

You can pass a list of window sizes (e.g. c(50000,100000,500000)) to automatically generate multiple views of each locus. This can even be a mix of different style inputs: e.g. c("1x","4.5x",25000).

verbose

Whether echolocatoR should be verbose or silent.

See also

Examples

data("BST1"); data("locus_dir");
track.Nott_histo <- NOTT_2019.epigenomic_histograms(finemap_dat = BST1, locus_dir = locus_dir, save_plot=F, return_assay_track=T, save_annot=F)
#> [1] "NOTT_2019:: Creating epigenomic histograms plot"
#> [1] "NOTT_2019:: Importing bigWig subsets from UCSC..."
#> [1] "+ Inferring genomic limits for window: 1x"
#> [1] "++ NOTT_2019:: Downloading and merging 12 peaks BED files."
#> Warning: scheduled core 1 encountered error in user code, all values of the job will be affected
#> Error in data.table::rbindlist(.): Item 1 of input is not a data.frame, data.table or list