Query and plot Roadmap epigenomic annotations

Query and plot Roadmap epigenomic annotations

ROADMAP.query_and_plot(
  subset_DT,
  results_path = "./ROADMAP",
  n_top_tissues = NULL,
  keyword_query = NULL,
  adjust = 0.2,
  force_new_query = F,
  remove_tmps = T
)

Arguments

subset_DT
Data.frame with at least the following columns:

SNP

SNP RSID

CHR

chromosome

POS

position

results_path
Where to store query results.

n_top_tissues
The number of top tissues to include, sorted by greatest number of rows (i.e. the number of genomic ranges within the window).

keyword_query
Search all columns in the Roadmap annotations metadata and only query annotations that contain your keywords. Can provide multiple keywords in list form: c("placenta","liver","monocytes")

force_new_query
Force a new query from the XGR database.

Value

A named list containing:

ggbio plot

GRanges object within the queried coordinates

See also

Examples

data("BST1")
finemap_DT <- BST1
roadmap_plot_query <- ROADMAP.query_and_plot(subset_DT=finemap_DT, keyword_query="monocytes")

#> [1] "ROADMAP:: Converting data to GRanges..."
#> [1] "+ ROADMAP:: 2 annotation(s) identified that match `keyword_query`."
#> [1] "+ Querying subset from Roadmap API: E124  -  1 / 2"
#> [1] "++ Downloading Roadmap Chromatin Marks: E124"
#> [1] "cd ./ROADMAP && tabix -p bed --begin 2 --end 3 https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/E124_15_coreMarks_dense.bed.bgz chr4:14737349-16737284"

#> Warning: File '/var/folders/sw/km9819713937hp640lj59ygw0000gn/T//RtmpuHIsxp/file124f0332a6b19' has size 0. Returning a NULL data.table.

#> Error in `[.data.frame`(x, i, j) : undefined columns selected
#> [1] "+ Querying subset from Roadmap API: E029  -  2 / 2"
#> [1] "++ Downloading Roadmap Chromatin Marks: E029"
#> [1] "cd ./ROADMAP && tabix -p bed --begin 2 --end 3 https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/E029_15_coreMarks_dense.bed.bgz chr4:14737349-16737284"

#> Warning: File '/var/folders/sw/km9819713937hp640lj59ygw0000gn/T//RtmpuHIsxp/file124f025dd508d' has size 0. Returning a NULL data.table.

#> Error in `[.data.frame`(x, i, j) : undefined columns selected
#> [1] "All downloads complete in 0 minutes"

#> Error in save_annotations(gr = grl.roadmap, anno_path = anno_path, libName = lib): could not find function "save_annotations"