Fine-map with SUSIE

Sum of Single Effects (SuSiE): Iterative Bayesian Step-wise Selection

SUSIE(
  subset_DT,
  LD_matrix,
  dataset_type = "GWAS",
  max_causal = 5,
  sample_size = NULL,
  prior_weights = NULL,
  PP_threshold = 0.95,
  scaled_prior_variance = 0.001,
  estimate_residual_variance = F,
  estimate_prior_variance = T,
  residual_variance = NULL,
  max_iter = 100,
  estimate_prior_method = "optim",
  manual_var_y = F,
  rescale_priors = T,
  plot_track_fit = F,
  return_all_CS = T,
  verbose = T
)

Arguments

dataset_type
The kind dataset you're fine-mapping (e.g. GWAS, eQTL, tQTL). This will also be used when creating the subdirectory where your results will be stored (e.g. Data/<dataset_type>/Kunkle_2019).

max_causal
The maximum number of non-zero effects (and thus causal variants).

sample_size
The overall sample size of the study. If none is given, and N_cases and N_controls columns are present, then sample_size is inferred to be: max(N_cases) + max(N_controls).

prior_weights
a p vector of prior probability that each element is non-zero

PP_threshold
The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set. For example, PP_threshold=.95 means that all Credible Set SNPs will be 95% Credible Set SNPs.

scaled_prior_variance
the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times)

estimate_residual_variance
indicates whether to estimate residual variance

estimate_prior_variance
indicates whether to estimate prior

residual_variance
the residual variance (defaults to variance of y)

max_iter
maximum number of iterations to perform

estimate_prior_method
The method used for estimating prior variance.

rescale_priors
If prior probabiltities are supplied, rescale them from 0-1 (i.e. rescaled_priors = priors / sum(priors)).

plot_track_fit
Record each iteration and make a GIF of the fine-mapping algorithm learning the causal variants. WARNING!: Making this plot can take a long time if there's many iterations.

verbose
if true outputs some progress messages

Source

GitHub Publication

Details

Notes on convergence: susieR will often give the warning: IBSS algorithm did not converge in 100 iterations!. This means the results might not necessarily be reliable. There's several things you can try to avoid this:

Make sure susieR is up-to-date: devtools::install_github("stephenslab/susieR@0.9.0")

Increase max_causal (e.g. 5 => 10).

Increase max_iter (e.g. 100 => 1000), though this will take longer.

Decrease the locus window size, which will also speed up the algorithm but potentially miss causal variants far from the lead SNP.

Changing estimate_prior_method does not seem to affect covergence warnings.

Notes on variance: GitHub Issue If estimate_residual_variance=TRUE without providing var_y and L>1, susieR will throw error: Estimating residual variance failed: the estimated value is negative Running susieR with var_y = var(b) provides exactly the same results.

Examples

data("BST1"); data("LD_matrix");
# LD_matrix <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/BST1/plink/UKB_LD.RDS")
finemap_DT <- SUSIE(subset_DT=BST1, LD_matrix=LD_matrix, estimate_residual_variance=T)

#> [1] "+ Preparing sample_size (N) column"
#> [1] "++ Computing effective sample size."
#> [1] "+ SUSIE:: sample_size= 216621"
#> [1] "+ SUSIE:: max_causal = 5"
#> [1] "+ Subsetting LD matrix and finemap_dat to common SNPs..."
#> [1] "+ LD:: Removing unnamed rows/cols"
#> [1] "+ LD:: Replacing NAs with 0"
#> [1] "+ LD_matrix = 95 SNPs."
#> [1] "+ finemap_dat = 6216 SNPs."
#> [1] "+ 95 SNPs in common."
#> [1] "+ SUSIE:: Using `susie_suff_stat()` from susieR v0.9.0"
#> [1] "+ SUSIE:: Extracting Credible Sets..."