Covert and query

If it is not tabix format already (determined by checking for a .tbi file of the same name in the same directory), the full summary statistics file is converted into tabix format for super fast querying. A query is then made using the min/max genomic positions to extract a locus-specific summary stats file.

TABIX(
  fullSS_path,
  study_dir = NULL,
  subset_path = NULL,
  is_tabix = F,
  chrom_col = "CHR",
  chrom_type = NULL,
  position_col = "POS",
  min_POS,
  max_POS,
  chrom,
  save_subset = T,
  nThread = 1,
  conda_env = "echoR",
  verbose = T
)

Arguments

fullSS_path
Path to the full summary statistics file (GWAS or QTL) that you want to fine-map. It is usually best to provide the absolute path rather than the relative path.

chrom_col
Name of the chromosome column in the full summary stats file. Can be "chr1" or "1" format. (default: ="CHR")

position_col
Name of the genomic position column in the full summary stats file. Must be in units of basepairs. (default: ="POS")

min_POS
Manually set the minimum genomic position for your locus subset. min_POS can clip the window size set by bp_distance. Can also be a list of positions (one for each locus) (e.g. min_POS=top_SNPs$min_POS).

max_POS
Manually set the maximum genomic position for your locus subset. max_POS can clip the window size set by bp_distance. Can also be a list of positions (one for each locus) (e.g. max_POS=top_SNPs$max_POS).

conda_env
The name of a conda environment to use.

verbose
Whether echolocatoR should be verbose or silent.

Value

data.table of locus subset summary statistics

See also

Examples

if (FALSE) {
data("locus_dir"); data("Nalls_top_SNPs")
fullSS_path <- "./results/GWAS/Nalls23andMe_2019/_genome_wide/nallsEtAl2019_allSamples_allVariants.mod.txt.gz"
top_SNPs <- import_topSNPs(topSS = Nalls_top_SNPs, chrom_col = "CHR", position_col = "BP", snp_col="SNP", pval_col="P, all studies", effect_col="Beta, all studies", gene_col="Nearest Gene", group_by_locus = T,locus_col = "Nearest Gene")
top_SNPs_BST1 <- subset(top_SNPs, Locus=='BST1')
bp_distance <- 1e+06
min_POS <- top_SNPs_BST1$POS - bp_distance
max_POS <- top_SNPs_BST1$POS + bp_distance
subset_file <- file.path(results_path,"BST1_Nalls23andMe_2019_subset.tsv.gz")
dat <- TABIX(fullSS_path=fullSS_path, subset_path="auto", min_POS=min_POS, max_POS=max_POS, chrom=top_SNPs_BST1$CHR)
}