Check overlap with XGR annotations

Automatically handles different file formats provided by XGR (e.g. varying kinds of nested/unnested GRanges). Then returns a Granges object with only the XGR annotation ranges that overlap with the SNPs in subset_DT. The GRanges merges hits from subset_DT.

XGR.iterate_overlap(
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes", "TFBS_Conserved",
    "ReMap_PublicAndEncode_TFBS", "Uniform_TFBS"),
  subset_DT,
  save_path = F,
  nCores = 4
)

Arguments

lib.selections
Which XGR annotations to check overlap with. For full list of libraries see: http://xgr.r-forge.r-project.org/#annotations-at-the-genomic-region-level

subset_DT
Data.frame with at least the following columns:

SNP

SNP RSID

CHR

chromosome

POS

position

save_path
Save the results as a data.frame

nCores
Multi-thread across libraries

See also

Examples

data("BST1")
finemap_DT <- BST1
gr.hits <- XGR.iterate_overlap(lib.selections=c("ENCODE_TFBS_ClusteredV3_CellTypes"), subset_DT=finemap_DT, nCores=1)

#> Start at 2020-07-09 21:30:14

#> 

#> 'ENCODE_TFBS_ClusteredV3_CellTypes' (from http://galahad.well.ox.ac.uk/bigdata/ENCODE_TFBS_ClusteredV3_CellTypes.RData) has been loaded into the working environment (at 2020-07-09 21:30:27)

#> 
#> End at 2020-07-09 21:30:27

#> Runtime in total is: 13 secs

#> [1] "+ dat2 already in GRanges format"

#> Warning: found more than one best sequence renaming map compatible with seqname style "NCBI" for this object, using the first one

#> 8702 query SNP(s) detected with reference overlap.

#> [1] "+ dat2 already in GRanges format"

#> Warning: found more than one best sequence renaming map compatible with seqname style "NCBI" for this object, using the first one

#> 25970 query SNP(s) detected with reference overlap.