Install tricky R packages

Some of echolocatoR's optional R package dependencies are especially tricky to install. Rather than including them in the DESCRIPTION file, which would require them to install echolocatoR at all, this functions installs them afterwards. Only packages not already installed will be installed.

extra_installs(
  cran_packages = F,
  xgr = T,
  bioc_packages = T,
  github_packages = T
)

Details

Some of the main packages installed via this function include:

gaston

CRAN

plotly

CRAN

XGR

xgr

foreign

xgr

refGenome

xgr

Rgraphviz

Bioconductor

biomaRt

Bioconductor

knitrBootstrap

GitHub

susieR

GitHub

Examples

library(echolocatoR)
extra_installs()

#> [1] "Installing additional echolocatoR R dependencies"

#> + Attempt 1: Installing XGR and deps via Bioconductor...

#> + Attempt 2: Installing XGR and deps via CRAN...

#> + Installing Bioconductor packages...

#> + Installing GitHub packages...

#> Downloading GitHub repo jimhester/knitrBootstrap@HEAD

#> 
#>   ✓  checking for file ‘/private/var/folders/zq/h7mtybc533b1qzkys_ttgpth0000gn/T/RtmpwfvQSD/remotes523e25c1898f/jimhester-knitrBootstrap-0f8d612/DESCRIPTION’
#>   ─  preparing ‘knitrBootstrap’:
#> ✓  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘knitrBootstrap_1.0.2.tar.gz’
#>      
#> 

#> Installing package into '/private/var/folders/zq/h7mtybc533b1qzkys_ttgpth0000gn/T/RtmplIUggK/temp_libpath50a46ff44e71'
#> (as 'lib' is unspecified)