Fine-map multiple loci

echolocatoR will automatically fine-map each locus. Uses the top_SNPs data.frame to define locus coordinates.

finemap_loci(
  loci,
  fullSS_path,
  dataset_name = "dataset_name",
  dataset_type = "dataset_type",
  force_new_subset = F,
  force_new_LD = F,
  force_new_finemap = T,
  results_dir = "./results",
  top_SNPs = "auto",
  finemap_methods = c("ABF", "SUSIE", "FINEMAP"),
  bp_distance = 5e+05,
  n_causal = 5,
  sample_size = NA,
  chrom_col = "CHR",
  position_col = "POS",
  snp_col = "SNP",
  pval_col = "P",
  effect_col = "Effect",
  stderr_col = "StdErr",
  tstat_col = "t-stat",
  MAF_col = "MAF",
  locus_col = "Locus",
  freq_col = "Freq",
  A1_col = "A1",
  A2_col = "A2",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  N_cases = NULL,
  N_controls = NULL,
  proportion_cases = "calculate",
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  LD_download_method = "direct",
  topVariants = 3,
  download_reference = T,
  min_POS = NA,
  max_POS = NA,
  min_MAF = NA,
  trim_gene_limits = F,
  max_snps = NULL,
  file_sep = "\t",
  min_r2 = 0,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  query_by = "coordinates",
  remove_variants = F,
  remove_correlates = F,
  probe_path = "./Data/eQTL/gene.ILMN.map",
  conditioned_snps = "auto",
  plot_LD = F,
  verbose = T,
  remove_tmps = T,
  PAINTOR_QTL_datasets = NULL,
  server = F,
  PP_threshold = 0.95,
  plot.types = c("simple"),
  plot.window = NULL,
  plot.Nott_epigenome = F,
  plot.Nott_binwidth = 2500,
  plot.Nott_bigwig_dir = NULL,
  plot.XGR_libnames = NULL,
  plot.Roadmap = F,
  plot.Roadmap_query = NULL,
  conda_env = "echoR"
)

Arguments

loci
The list of loci you want to fine-map

fullSS_path
Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.

dataset_name
The name you want to assign to the dataset being fine-mapped, This will be used to name the subdirectory where your results will be stored (e.g. Data/GWAS/<dataset_name>). Don't use special characters (e.g.".", "/").

dataset_type
The kind dataset you're fine-mapping (e.g. GWAS, eQTL, tQTL). This will also be used when creating the subdirectory where your results will be stored (e.g. Data/<dataset_type>/Kunkle_2019).

force_new_subset
By default, if a subset of the full summary stats file for a given locus is already present, then echolocatoR will just use the preexisting file. Set force_new_subset=T to override this and extract a new subset. Subsets are saved in the following path structure: Data/<dataset_type>/<dataset_name>/<locus>/Multi-finemap/<locus>_<dataset_name>_Multi-finemap.tsv.gz

force_new_LD
By default, if an LD matrix file for a given locus is already present, then echolocatoR will just use the preexisting file. Set force_new_LD=T to override this and extract a new subset.

force_new_finemap
By default, if an fine-mapping results file for a given locus is already present, then echolocatoR will just use the preexisting file. Set force_new_finemap=T to override this and re-run fine-mapping.

results_dir
Where to store all results.

top_SNPs
A data.frame with the genomic coordinates of the lead SNP for each locus. The lead SNP will be used as the center of the window when extracting subset from the full GWAS/QTL summary statistics file. Only one SNP per Locus should be included. At minimum, top_SNPs should include the following columns:

Locus

A unique name for each locus. Often, loci are named after a relevant gene (e.g. LRRK2) or based on the name/coordinates of the lead SNP (e.g. locus_chr12_40734202)

CHR

The chromosome that the SNP is on. Can be "chr12" or "12" format.

POS

The genomic position of the SNP (in basepairs)

finemap_methods
Which fine-mapping methods you want to use.

bp_distance
The width of the window size you want each locus to be. For example, if bp_distance=500000 then the locus will span 500kb from the lead SNP in either direction, resulting in a locus that is ~1Mb long (depending on the dataset).

n_causal
The maximum number of potential causal SNPs per locus. This parameter is used somewhat differntly by different fine-mapping tools. See tool-specific functions for details.

sample_size
The overall sample size of the study. If none is given, and N_cases and N_controls columns are present, then sample_size is inferred to be: max(N_cases) + max(N_controls).

chrom_col
Name of the chromosome column in the full summary stats file. Can be "chr1" or "1" format. (default: ="CHR")

position_col
Name of the genomic position column in the full summary stats file. Must be in units of basepairs. (default: ="POS")

snp_col
Name of the SNP RSID column in the full summary stats file. (default: ="SNP")

pval_col
Name of the p-value column in the full summary stats file. Raw p-values are preferred, but if not available corrected p-values (e.g. FDR) can be used instead. (default: ="P")

effect_col
Name of the effect size column in the full summary stats file. Effect size is preferred, but if not available other metrics like Beta for Odds Ratio can be used instead. (default: ="Effect")

stderr_col
Name of the standard error column in the full summary stats file. You can also set stderr_col="calculate" to infer standard error using: effect / tstat. (default: ="StdErr")

tstat_col
Name of the t-statistic column in the full summary stats file. This column is not necessary unless stderr_col="calculate" or the standard error column is missing. (default: ="t-stat")

MAF_col
Name of the minor allele frequency column in the full summary stats file. Can be inferred from freq_col if missing from the dataset. (default: ="MAF")

locus_col
Name of the locus column in the full summary stats file. (default: ="Locus")

freq_col
Name of the allele frequency column in the full summary stats file. Effect allele frequency is preferred, but the non-effect allele can be provided instead (though this may be less accurate). This column is not necessary unless MAF_col="calculate" or the MAF column is missing. (default: ="Freq")

A1_col
Name of the effect/risk allele column in the full summary stats. IMPORTANT: Make sure this actually the case for your full summary stats file. Unfortunately, different studies report different kinds of allele information in a non-standardized way. Meaning that A1/A2 can refer to any number of things:

effect/other alleles

in the case of diseases

ref/alt alleles

where ref is the reference genome being used

major/minor alleles

This dichotomy holds true for bi-allelic SNPs but not necessary multi-allelic SNPs

This makes comparing summary stats across GWAS/QTL datasets very confusing for several reasons:

Multi-allelic SNPs

SNPs can have more than just 2 possible alleles (multi-allelic SNPs). Even if you compare the same SNP between two studies, you may accidentally be comparing totally different alleles.

Valence

The valence (+/-) of per-SNP GWAS effect sizes/beta can be relative to different allele types between studies. For example, let's say in one GWAS study your effect size for SNP A is 1.5 relative to the major allele in one study, and the minor allele happens to be the one found in the reference genome. You then try to compare that effect size to that of the same SNP in another GWAS. But, the valence of the effect sizes in the 2nd GWAS study are all relative to the reference genome (instead of the minor allele), giving the same SNP a value of -1.2. If you took the effect sizes at face value you'd say the signals are in opposite directions. But once you take into account how the valences were determined in each study you realize that they're actually both positive relative to the major allele.

This process of reversing per-SNP valences based on aligning the alleles is known as allele flipping. This is important when comparing individual SNPs, but can also have an impact on colocalization results.

N_cases_col
Name of the column in the full summary stats that has the number of case subjects in the study. This can either be per SNP sample sizes, or one number repeated across all rows. Proxy cases (e.g. relatives of people with the disease being investigated) should be included in this estimate if any were used in the study. This column is not necesssary if N_cases parameter is provided. (default: ="N_cases")

N_controls_col
Name of the column in the full summary stats that has the number of control subjects in the study. This can either be per SNP sample sizes, or one number repeated across all rows. This column is not necesssary if N_controls parameter is provided. (default: ="N_controls")

N_cases
The number of case subjects in the study. Instead of providing a redundant N_cases_col column, you can simply enter one value here.

N_controls
The number of control subjects in the study. Instead of providing a redundant N_controls_col column, you can simply enter one value here.

proportion_cases
The proportion of total subjects in the study that were cases. if proportion_cases="calculate" then this is inferred: N_controls / N_controls.

download_reference
When acquiring LD matrixes, the default is to delete the full vcf or npz files after echolocator has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set download_reference=T.

min_POS
Manually set the minimum genomic position for your locus subset. min_POS can clip the window size set by bp_distance. Only use this parameter when fine-mapping one locus at a time.

max_POS
Manually set the maximum genomic position for your locus subset. max_POS can clip the window size set by bp_distance. Only use this parameter when fine-mapping one locus at a time.

min_MAF
Remove any SNPs with MAF < min_MAF.

trim_gene_limits
If a valid gene symbol is provided to trim_gene_limits, the gene's canonical coordinates are pulled from biomaRt. This includes introns, exons, and proximal regulatory regions (e.g. promoters). Any SNPs that fall outside these coordinates are remove from downstream fine-mapping. Set trim_gene_limits=F to not limit by gene coordinates (default).

max_snps
The maximum number of SNPs to include in the locus. If the current window size yields > max_snps, then the outer edges of the of the locus are trimmed until the number of SNPs ≤ max_snps.

file_sep
The separator in the full summary stats file. This parameter is only necessary if query_by!="tabix".

min_r2
Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.

LD_block
Calculate LD blocks with plink and only include the block to which the lead SNP belongs.

LD_block_size
Adjust the granularity of block sizes when LD_block=T.

min_Dprime
Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus. This is paramter currently only works when LD_reference!="UKB".

query_by
Choose which method you want to use to extract locus subsets from the full summary stats file. Methods include:

"tabix"

Convert the full summary stats file in an indexed tabix file. Makes querying lightning fast after the initial conversion is done. (default)

"coordinates"

Extract locus subsets using min/max genomic coordinates with awk.

remove_variants
A list of variants to remove from the locus subset file.

remove_correlates
If remove_correlates is set to a value between 0-1, removes any SNPs that are in LD with any of the remove_variants above the threshold provided by remove_correlates.

probe_path
The location of the file containing translations between probe IDs and gene symbols. Only used for certain eQTL datasets.

conditioned_snps
Which SNPs to conditions on when fine-mapping with COJO.

plot_LD
Whether to plot a subset of the LD matix.

verbose
Whether echolocatoR should be verbose or silent.

remove_tmps
Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.

PAINTOR_QTL_datasets
A list of QTL datasets to be used when conducting joint functional fine-mapping with PAINTOR.

server
Whether echolocatoR is being run on a computing cluster/server or on a local machine.

PP_threshold
The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set. For example, PP_threshold=.95 means that all Credible Set SNPs will be 95% Credible Set SNPs.

plot.types
Which kinds of plots to include. Options:

"simple"

Just plot the following tracks: GWAS, fine-mapping, gene models

"fancy"

Additionally plot XGR annotation tracks (XGR, Roadmap, Nott).

plot.window
Zoom into the center of the locus when plotting (without editing the fine-mapping results file).

plot.Nott_binwidth
When including Nott et al. (2019) epigenomic data in the track plots, adjust the bin width of the histograms.

plot.Nott_bigwig_dir
Instead of pulling Nott et al. (2019) epigenomic data from the UCSC Genome Browsers, use a set of local bigwig files.

plot.Roadmap
Find and plot annotations from Roadmap.

plot.Roadmap_query
Only plot annotations from Roadmap whose metadata contains a string or any items from a list of strings (e.g. "brain" or c("brain","liver","monocytes"))

conda_env
The name of a conda environment to use.

subset_path
The file you want your locus subset saved as. Only use when fine-mapping one locus at a time. If subset_path="auto" (default), a locus subset file name is automatically constructed as: Data/<dataset_type>/<dataset_name>/<locus>/Multi-finemap/<locus>_<dataset_name>_Multi-finemap.tsv.gz

Value

A merged data.frame with all fine-mapping results from all loci.

See also