Merge fine-mapping results from all loci

Gather fine-mapping results from echolocatoR across all loci and merge into a single data.frame.

merge_finemapping_results(
  dataset = "./Data/GWAS",
  minimum_support = 0,
  include_leadSNPs = T,
  xlsx_path = F,
  from_storage = T,
  haploreg_annotation = F,
  regulomeDB_annotation = F,
  biomart_annotation = F,
  PP_threshold = 0.95,
  consensus_thresh = 2,
  exclude_methods = NULL,
  verbose = T,
  nThread = 4
)

Arguments

dataset
Path to the folder you want to recursively search for results files within (e.g. "Data/GWAS/Nalls23andMe_2019"). Set this to a path that includes multiple subfolders if you want to gather results from multiple studies at once (e.g. "Data/GWAS").

minimum_support
Filter SNPs by the minimum number of fine-mapping tools that contained the SNP in their Credible Set.

include_leadSNPs
Include lead GWAS/QTL SNPs per locus (regardless of other filtering criterion).

xlsx_path
Save merged data.frame as excel file.

from_storage
Search for stored results files.

haploreg_annotation
Annotate SNPs with HaploReg (using HaploR).

regulomeDB_annotation
Annotate SNPs with regulaomeDB (using HaploR).

biomart_annotation
Annotate SNPs with biomart.

PP_threshold
Mean posterior probability threshold to include SNPs in mean PP Credible Set (averaged across all fine-mapping tools).

consensus_thresh
The minimum number of tools that have the SNPs in their Credible Set to classify it as a Consensus_SNP.

exclude_methods
Exclude certain fine-mapping methods when estimating mean.CS and Consensus_SNP.

verbose
Print messages.

Examples

dataset_dir <- "~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"
# UCS and lead SNPs: No annotation
merged_DT <- merge_finemapping_results(dataset=dataset_dir, minimum_support=1, include_leadSNPs=T)

#> [1] "+ Gathering all fine-mapping results from storage..."
#> [1] "+ Importing results... ASXL3"
#> [1] "+ Importing results... ATG14"
#> [1] "+ Importing results... ATP6V0A1"
#> [1] "+ Importing results... BIN3"
#> [1] "+ Importing results... BRIP1"
#> [1] "+ Importing results... BST1"
#> [1] "+ Importing results... C5orf24"
#> [1] "+ Importing results... CAB39L"
#> [1] "+ Importing results... CAMK2D"
#> [1] "+ Importing results... CASC16"
#> [1] "+ Importing results... CD19"
#> [1] "+ Importing results... CHD9"
#> [1] "+ Importing results... CHRNB1"
#> [1] "+ Importing results... CLCN3"
#> [1] "+ Importing results... CRHR1"
#> [1] "+ Importing results... CRLS1"
#> [1] "+ Importing results... CTSB"
#> [1] "+ Importing results... DLG2"
#> [1] "+ Importing results... DNAH17"
#> [1] "+ Importing results... DYRK1A"
#> [1] "+ Importing results... ELOVL7"
#> [1] "+ Importing results... FAM171A2"
#> [1] "+ Importing results... FAM47E-STBD1"
#> [1] "+ Importing results... FAM49B"
#> [1] "+ Importing results... FBRSL1"
#> [1] "+ Importing results... FCGR2A"
#> [1] "+ Importing results... FGF20"
#> [1] "+ Importing results... FYN"
#> [1] "+ Importing results... GALC"
#> [1] "+ Importing results... GBF1"
#> [1] "+ Importing results... GCH1"
#> [1] "+ Importing results... GPNMB"
#> [1] "+ Importing results... GS1-124K5.11"
#> [1] "+ Importing results... HIP1R"
#> [1] "+ Importing results... HLA-DRB5"
#> [1] "+ Importing results... IGSF9B"
#> [1] "+ Importing results... INPP5F"
#> [1] "+ Importing results... IP6K2"
#> [1] "+ Importing results... ITGA8"
#> [1] "+ Importing results... ITPKB"
#> [1] "+ Importing results... KCNIP3"
#> [1] "+ Importing results... KCNS3"
#> [1] "+ Importing results... KPNA1"
#> [1] "+ Importing results... KRTCAP2"
#> [1] "+ Importing results... LCORL"
#> [1] "+ Importing results... LINC00693"
#> [1] "+ Importing results... LMNB1"
#> [1] "+ Importing results... LOC100131289"
#> [1] "+ Importing results... LRRK2"
#> [1] "+ Importing results... MAP4K4"
#> [1] "+ Importing results... MBNL2"
#> [1] "+ Importing results... MCCC1"
#> [1] "+ Importing results... MED12L"
#> [1] "+ Importing results... MEX3C"
#> [1] "+ Importing results... MIPOL1"
#> [1] "+ Importing results... NOD2"
#> [1] "+ Importing results... NUCKS1"
#> [1] "+ Importing results... PAM"
#> [1] "+ Importing results... RETREG3"
#> [1] "+ Importing results... RIMS1"
#> [1] "+ Importing results... RIT2"
#> [1] "+ Importing results... RNF141"
#> [1] "+ Importing results... RPS12"
#> [1] "+ Importing results... RPS6KL1"
#> [1] "+ Importing results... SATB1"
#> [1] "+ Importing results... SCAF11"
#> [1] "+ Importing results... SETD1A"
#> [1] "+ Importing results... SH3GL2"
#> [1] "+ Importing results... SIPA1L2"
#> [1] "+ Importing results... SLC2A13"
#> [1] "+ Importing results... SNCA"
#> [1] "+ Importing results... SP1"
#> [1] "+ Importing results... SPPL2B"
#> [1] "+ Importing results... SPTSSB"
#> [1] "+ Importing results... STK39"
#> [1] "+ Importing results... SYT17"
#> [1] "+ Importing results... TMEM163"
#> [1] "+ Importing results... TMEM175"
#> [1] "+ Importing results... TRIM40"
#> [1] "+ Importing results... UBAP2"
#> [1] "+ Importing results... VAMP4"
#> [1] "+ Importing results... VPS13C"
#> [1] "+ Importing results... WNT3"
#> [1] "+ Identifying Consensus SNPs..."
#> [1] "++ support_thresh = 2"
#> [1] "+ Calculating mean Posterior Probability (mean.PP)..."
#> [1] "++ 4 fine-mapping methods used."
#> [1] "++ 680 Credible Set SNPs identified."
#> [1] "++ 224 Consensus SNPs identified."

# UCS and lead SNPs: With annotations
merged_DT <- merge_finemapping_results(dataset=dataset_dir, minimum_support=1, include_leadSNPs=T, haploreg_annotation=T, biomart_annotation=T)

#> [1] "+ Gathering all fine-mapping results from storage..."
#> [1] "+ Importing results... ASXL3"
#> [1] "+ Importing results... ATG14"
#> [1] "+ Importing results... ATP6V0A1"
#> [1] "+ Importing results... BIN3"
#> [1] "+ Importing results... BRIP1"
#> [1] "+ Importing results... BST1"
#> [1] "+ Importing results... C5orf24"
#> [1] "+ Importing results... CAB39L"
#> [1] "+ Importing results... CAMK2D"
#> [1] "+ Importing results... CASC16"
#> [1] "+ Importing results... CD19"
#> [1] "+ Importing results... CHD9"
#> [1] "+ Importing results... CHRNB1"
#> [1] "+ Importing results... CLCN3"
#> [1] "+ Importing results... CRHR1"
#> [1] "+ Importing results... CRLS1"
#> [1] "+ Importing results... CTSB"
#> [1] "+ Importing results... DLG2"
#> [1] "+ Importing results... DNAH17"
#> [1] "+ Importing results... DYRK1A"
#> [1] "+ Importing results... ELOVL7"
#> [1] "+ Importing results... FAM171A2"
#> [1] "+ Importing results... FAM47E-STBD1"
#> [1] "+ Importing results... FAM49B"
#> [1] "+ Importing results... FBRSL1"
#> [1] "+ Importing results... FCGR2A"
#> [1] "+ Importing results... FGF20"
#> [1] "+ Importing results... FYN"
#> [1] "+ Importing results... GALC"
#> [1] "+ Importing results... GBF1"
#> [1] "+ Importing results... GCH1"
#> [1] "+ Importing results... GPNMB"
#> [1] "+ Importing results... GS1-124K5.11"
#> [1] "+ Importing results... HIP1R"
#> [1] "+ Importing results... HLA-DRB5"
#> [1] "+ Importing results... IGSF9B"
#> [1] "+ Importing results... INPP5F"
#> [1] "+ Importing results... IP6K2"
#> [1] "+ Importing results... ITGA8"
#> [1] "+ Importing results... ITPKB"
#> [1] "+ Importing results... KCNIP3"
#> [1] "+ Importing results... KCNS3"
#> [1] "+ Importing results... KPNA1"
#> [1] "+ Importing results... KRTCAP2"
#> [1] "+ Importing results... LCORL"
#> [1] "+ Importing results... LINC00693"
#> [1] "+ Importing results... LMNB1"
#> [1] "+ Importing results... LOC100131289"
#> [1] "+ Importing results... LRRK2"
#> [1] "+ Importing results... MAP4K4"
#> [1] "+ Importing results... MBNL2"
#> [1] "+ Importing results... MCCC1"
#> [1] "+ Importing results... MED12L"
#> [1] "+ Importing results... MEX3C"
#> [1] "+ Importing results... MIPOL1"
#> [1] "+ Importing results... NOD2"
#> [1] "+ Importing results... NUCKS1"
#> [1] "+ Importing results... PAM"
#> [1] "+ Importing results... RETREG3"
#> [1] "+ Importing results... RIMS1"
#> [1] "+ Importing results... RIT2"
#> [1] "+ Importing results... RNF141"
#> [1] "+ Importing results... RPS12"
#> [1] "+ Importing results... RPS6KL1"
#> [1] "+ Importing results... SATB1"
#> [1] "+ Importing results... SCAF11"
#> [1] "+ Importing results... SETD1A"
#> [1] "+ Importing results... SH3GL2"
#> [1] "+ Importing results... SIPA1L2"
#> [1] "+ Importing results... SLC2A13"
#> [1] "+ Importing results... SNCA"
#> [1] "+ Importing results... SP1"
#> [1] "+ Importing results... SPPL2B"
#> [1] "+ Importing results... SPTSSB"
#> [1] "+ Importing results... STK39"
#> [1] "+ Importing results... SYT17"
#> [1] "+ Importing results... TMEM163"
#> [1] "+ Importing results... TMEM175"
#> [1] "+ Importing results... TRIM40"
#> [1] "+ Importing results... UBAP2"
#> [1] "+ Importing results... VAMP4"
#> [1] "+ Importing results... VPS13C"
#> [1] "+ Importing results... WNT3"
#> [1] "+ Identifying Consensus SNPs..."
#> [1] "++ support_thresh = 2"
#> [1] "+ Calculating mean Posterior Probability (mean.PP)..."
#> [1] "++ 4 fine-mapping methods used."
#> [1] "++ 680 Credible Set SNPs identified."
#> [1] "++ 224 Consensus SNPs identified."
#> [1] "+ Gathering annotation data from HaploReg..."
#> [1] "++ Submitting chunk 1 / 1"
#> [1] "+ Gathering annotation data from Biomart..."

#> Cache found