Fine-map using multiple fine-mapping tools

Fine-mapping will be repeated on the same locus using each of the tools in finemap_method_list. Then, all results will be merged into the locus-specific multi-finemap file, along with the original per-SNP GWAS/QTL summary statistics. Each tools will have the following columns:

<tool>.PP

The posterior probability (PP) of a SNP being causal for the trait. Though this is a generalization and the exact meaning of PP will differ by tools (e.g. Posterior Inclusion Probability for SUSIE).

<tool>.CS

Which credible set the SNP is part of (within a locus). If =0, then the SNP was not part of any credible set. Some tools only produce one credible set per locus.

multi_finemap(
  locus_dir,
  fullSS_path,
  finemap_method_list,
  finemap_args = NULL,
  subset_DT,
  dataset_type,
  LD_matrix = NULL,
  n_causal = 5,
  sample_size = NULL,
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  case_control = T,
  verbose = T,
  nThread = 4,
  conda_env = "echoR"
)

Arguments

fullSS_path

Path to the full summary statistics file (GWAS or QTL) that you want to fine-map. It is usually best to provide the absolute path rather than the relative path.

dataset_type

The kind dataset you're fine-mapping (e.g. GWAS, eQTL, tQTL). This will also be used when creating the subdirectory where your results will be stored (e.g. Data/<dataset_type>/Kunkle_2019).

n_causal

The maximum number of potential causal SNPs per locus. This parameter is used somewhat differntly by different fine-mapping tools. See tool-specific functions for details.

sample_size

The overall sample size of the study. If none is given, and N_cases and N_controls columns are present, then sample_size is inferred to be: max(N_cases) + max(N_controls).

PAINTOR_QTL_datasets

A list of QTL datasets to be used when conducting joint functional fine-mapping with PAINTOR.

PP_threshold

The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set. For example, PP_threshold=.95 means that all Credible Set SNPs will be 95% Credible Set SNPs.

verbose

Whether echolocatoR should be verbose or silent.

conda_env

The name of a conda environment to use.

See also

Examples

if (FALSE) {
data("BST1"); data("BST1_LD_matrix");
subset_DT <- BST1
finemap_method_list <- c("ABF","SUSIE")
}