Use awk to query locus subsets.

Search full summary stats file by genomics coordinates. To be used when CHR and POS are merged into one column (e.g. chr12:12209944).

query_by_coordinates_merged(
  top_SNPs,
  fullSS_path,
  subset_path,
  locus,
  chrom_col,
  file_sep = " ",
  location_sep = ":",
  min_POS,
  max_POS,
  bp_distance
)

Arguments

top_SNPs
A data.frame with the genomic coordinates of the lead SNP for each locus. The lead SNP will be used as the center of the window when extracting subset from the full GWAS/QTL summary statistics file. Only one SNP per Locus should be included. At minimum, top_SNPs should include the following columns:

Locus

A unique name for each locus. Often, loci are named after a relevant gene (e.g. LRRK2) or based on the name/coordinates of the lead SNP (e.g. locus_chr12_40734202)

CHR

The chromosome that the SNP is on. Can be "chr12" or "12" format.

POS

The genomic position of the SNP (in basepairs)

fullSS_path
Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.

chrom_col
Name of the chromosome column in the full summary stats file. Can be "chr1" or "1" format. (default: ="CHR")

file_sep
The separator in the full summary stats file. This parameter is only necessary if query_by!="tabix".

location_sep
The separator character when CHR and POS are merged into one column (e.g. ":" when formatted like chr12:12209944).

min_POS
Manually set the minimum genomic position for your locus subset. min_POS can clip the window size set by bp_distance. Only use this parameter when fine-mapping one locus at a time.

max_POS
Manually set the maximum genomic position for your locus subset. max_POS can clip the window size set by bp_distance. Only use this parameter when fine-mapping one locus at a time.

bp_distance
The width of the window size you want each locus to be. For example, if bp_distance=500000 then the locus will span 500kb from the lead SNP in either direction, resulting in a locus that is ~1Mb long (depending on the dataset).

See also