Use awk to query locus subsets.

To be used when probe name (but not gene name) is a column in the full summary stats file. Requires a gene-probe key file (probe_path). More commonly useful for QTL full summary stats files.

query_by_probe(
  fullSS_path,
  subset_path,
  gene,
  gene_col,
  chrom_col,
  file_sep,
  probe_path,
  coordinates_merged = T,
  location_sep = ":"
)

Arguments

fullSS_path
Path to the full summary statistics file (GWAS or QTL) that you want to fine-map. It is usually best to provide the absolute path rather than the relative path.

gene_col
For QTL studies, the name of the [e]gene column in the full summary stats file (default: "gene"). This column will be used for filtering summary stats if supplying a named list of gene:Locus pairs to loci.

chrom_col
Name of the chromosome column in the full summary stats file. Can be "chr1" or "1" format. (default: ="CHR")

file_sep
The separator in the full summary stats file. This parameter is only necessary if query_by!="tabix".

probe_path
The location of the file containing translations between probe IDs and gene symbols. Only used for certain eQTL datasets.

coordinates_merged
Whether CHR and POS are merged into one column (e.g. chr12:12209944).

location_sep
The separator character when CHR and POS are merged into one column (e.g. ":" when formatted like chr12:12209944).

See also