Fine-map locus with FINEMAP

The stepwise conditional search starts with a causal configuration containing the SNP with the lowest P-value alone and then iteratively adds to the causal configuration the SNP given the highest posterior model probability until no further SNP yields a higher posterior model probability.

FINEMAP(
  subset_DT,
  locus_dir,
  LD_matrix,
  FINEMAP_path = NULL,
  n_samples = NULL,
  n_causal = 5,
  model = "cond",
  remove_tmps = T,
  credset_thresh = 0.95,
  finemap_version = "1.3.1",
  server = F
)

Source

http://www.christianbenner.com

See also

Examples

data("locus_dir"); data("BST1"); data("BST1_LD_matrix");
finemap_DT <- BST1
dir.create(file.path(locus_dir,"FINEMAP"), showWarnings = FALSE, recursive = TRUE)
out <- subset_common_snps(LD_matrix=LD_matrix, finemap_DT=finemap_DT)

#> Error in subset_common_snps(LD_matrix = LD_matrix, finemap_DT = finemap_DT): unused argument (finemap_DT = finemap_DT)

LD_matrix <- out$LD

#> Error in eval(expr, envir, enclos): object 'out' not found

finemap_DT <- out$DT

#> Error in eval(expr, envir, enclos): object 'out' not found

finemap_DT <- FINEMAP(subset_DT=finemap_DT, locus_dir=locus_dir, LD_matrix=LD_matrix)

#> Warning: no non-missing arguments to max; returning -Inf

#> [1] "++ Constructing FINEMAP master file."
#> [1] "++ Formatting data.z file for FINEMAP"
#> [1] "++ Formatting LD Matrix for FINEMAP"
#> [1] "+++ Writing FINEMAP z and ld files..."
#> [1] "cd results/GWAS/Nalls23andMe_2019/BST1 && /private/var/folders/sw/km9819713937hp640lj59ygw0000gn/T/RtmpQyU3yJ/temp_libpath121843faaf415/echolocatoR/tools/FINEMAP/finemap_v1.3_MacOSX --cond --in-files results/GWAS/Nalls23andMe_2019/BST1/FINEMAP/master --log --n-causal-snps 5"

#> Error in data.table::fread(file.path(locus_dir, "FINEMAP/data.config")): File 'results/GWAS/Nalls23andMe_2019/BST1/FINEMAP/data.config' does not exist or is non-readable. getwd()=='/Users/schilder/Desktop/echolocatoR/docs/reference'