Compute LD from 1000 Genomes

Downloads a subset vcf of the 1KG database that matches your locus coordinates. Then uses plink to calculate LD on the fly.

LD.1KG(
  locus_dir,
  subset_DT,
  locus = NULL,
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  vcf_folder = NULL,
  download_reference = T,
  min_r2 = F,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  remove_correlates = F,
  remove_tmps = T,
  fillNA = 0,
  nThread = 4,
  download_method = "wget",
  verbose = T
)

Arguments

download_reference
When acquiring LD matrixes, the default is to delete the full vcf or npz files after echolocatoR has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set download_reference=T.

min_r2
Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.

LD_block
Calculate LD blocks with plink and only include the block to which the lead SNP belongs.

LD_block_size
Adjust the granularity of block sizes when LD_block=T.

min_Dprime
Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus. This is paramter currently only works when LD_reference!="UKB".

remove_correlates
If remove_correlates is set to a value between 0-1, removes any SNPs that are in LD with any of the remove_variants above the threshold provided by remove_correlates.

@param LD_reference Which linkage disequilibrium reference panel do you want to use. Options include:

"UKB"

A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See Wiessbrod et al. (2019) for more details.

"1KGphase1"

Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.

"1KGphase3"

Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.

@param superpopulation Subset your LD reference panel by superopulation. Setting the superpopulation is not currently possible when LD_reference="UKB". 1KGphase1 options include:

"AFR"

African [descent]

"AMR"

Ad-mixed American

"EAS"

East Asian

"EUR"

European

"SAS"

South Asian

remove_tmps
Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.

fillNA
When pairwise LD (r) between two SNPs is NA, replace with 0.

verbose
Whether echolocatoR should be verbose or silent.

Details

This approach is taken, because other API query tools have limitations with the window size being queried. This approach does not have this limitations, allowing you to fine-map loci more completely.

See also

Examples

if (FALSE) {
data("BST1"); data("locus_dir");
LD_matrix <- LD.1KG(locus_dir=file.path("~/Desktop",locus_dir), subset_DT=BST1, locus="BST1", LD_reference="1KGphase1")
}