Procure an LD matrix for fine-mapping

Calculate and/or query linkage disequilibrium (LD) from reference panels (UK Biobank, 1000 Genomes), or user-supplied datasets.

LD.load_or_create(
  locus_dir,
  subset_DT,
  locus = NULL,
  force_new_LD = F,
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  download_reference = T,
  download_method = "direct",
  vcf_folder = NULL,
  min_r2 = 0,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0,
  verbose = T,
  server = F,
  remove_tmps = T,
  nThread = 4
)

Arguments

subset_DT
The locus subset of the full summary stats file.

force_new_LD
By default, if an LD matrix file for a given locus is already present, then echolocatoR will just use the preexisting file. Set force_new_LD=T to override this and extract a new subset.

download_reference
When acquiring LD matrixes, the default is to delete the full vcf or npz files after echolocator has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set download_reference=T.

min_r2
Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.

LD_block
Calculate LD blocks with plink and only include the block to which the lead SNP belongs.

LD_block_size
Adjust the granularity of block sizes when LD_block=T.

min_Dprime
Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus. This is paramter currently only works when LD_reference!="UKB".

remove_correlates
If remove_correlates is set to a value between 0-1, removes any SNPs that are in LD with any of the remove_variants above the threshold provided by remove_correlates.

verbose
Whether echolocatoR should be verbose or silent.

server
Whether echolocatoR is being run on a computing cluster/server or on a local machine.

remove_tmps
Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.

Value

A symmetric LD matrix of pairwise r values.

Details

Options:

Download pre-computed LD matrix from UK Biobank.

Download raw vcf file from 1KG and compute LD on the fly.

Compute LD fon the fly from a user-supplied vcf file.

Use a user-supplied pre-computed LD-matrix.

See also

Examples

if (FALSE) {
data("BST1"); data("locus_dir");
locus_dir <- file.path("~/Desktop",locus_dir)
# UK Biobank LD
LD.load_or_create(locus_dir=locus_dir, subset_DT=BST1, locus="BST1", )
}