Query and plot Roadmap epigenomic annotations

Query and plot Roadmap epigenomic annotations

ROADMAP.query_and_plot(
  subset_DT,
  results_path = "./ROADMAP",
  n_top_tissues = NULL,
  keyword_query = NULL,
  adjust = 0.2,
  force_new_query = F,
  remove_tmps = T
)

Arguments

subset_DT

Data.frame with at least the following columns:

SNP

SNP RSID

CHR

chromosome

POS

position

results_path

Where to store query results.

n_top_tissues

The number of top tissues to include, sorted by greatest number of rows (i.e. the number of genomic ranges within the window).

keyword_query

Search all columns in the Roadmap annotations metadata and only query annotations that contain your keywords. Can provide multiple keywords in list form: c("placenta","liver","monocytes")

force_new_query

Force a new query from the XGR database.

Value

A named list containing:

ggbio plot

GRanges object within the queried coordinates

See also

Examples

if (FALSE) {
data("BST1")
finemap_DT <- BST1
roadmap_plot_query <- ROADMAP.query_and_plot(subset_DT=finemap_DT, keyword_query="monocytes")
}