Fine-map with SUSIE

Sum of Single Effects (SuSiE): Iterative Bayesian Step-wise Selection

SUSIE(
  subset_DT,
  LD_matrix,
  dataset_type = "GWAS",
  max_causal = 5,
  sample_size = NULL,
  prior_weights = NULL,
  PP_threshold = 0.95,
  scaled_prior_variance = 0.001,
  estimate_residual_variance = F,
  estimate_prior_variance = T,
  residual_variance = NULL,
  max_iter = 100,
  estimate_prior_method = "optim",
  manual_var_y = F,
  rescale_priors = T,
  plot_track_fit = F,
  return_all_CS = T,
  verbose = T
)

Source

GitHub Publication

Arguments

dataset_type

The kind dataset you're fine-mapping (e.g. GWAS, eQTL, tQTL). This will also be used when creating the subdirectory where your results will be stored (e.g. Data/<dataset_type>/Kunkle_2019).

max_causal

The maximum number of non-zero effects (and thus causal variants).

sample_size

The overall sample size of the study. If none is given, and N_cases and N_controls columns are present, then sample_size is inferred to be: max(N_cases) + max(N_controls).

prior_weights

A vector of length p, in which each entry gives the prior probability that corresponding column of X has a nonzero effect on the outcome, y.

PP_threshold

The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set. For example, PP_threshold=.95 means that all Credible Set SNPs will be 95% Credible Set SNPs.

scaled_prior_variance

The prior variance, divided by var(y) (or by (1/(n-1))yty for susie_suff_stat); that is, the prior variance of each non-zero element of b is var(y) * scaled_prior_variance. The value provided should be either a scalar or a vector of length L. If estimate_prior_variance = TRUE, this provides initial estimates of the prior variances.

estimate_residual_variance

If estimate_residual_variance = TRUE, the residual variance is estimated, using residual_variance as an initial value. If estimate_residual_variance = FALSE, the residual variance is fixed to the value supplied by residual_variance.

estimate_prior_variance

If estimate_prior_variance = TRUE, the prior variance is estimated (this is a separate parameter for each of the L effects). If provided, scaled_prior_variance is then used as an initial value for the optimization. When estimate_prior_variance = FALSE, the prior variance for each of the L effects is determined by the value supplied to scaled_prior_variance.

residual_variance

Variance of the residual. If estimate_residual_variance = TRUE, this value provides the initial estimate of the residual variance. By default, it is set to var(y) in susie and (1/(n-1))yty in susie_suff_stat.

max_iter

Maximum number of IBSS iterations to perform.

estimate_prior_method

The method used for estimating prior variance. When estimate_prior_method = "simple" is used, the likelihood at the specified prior variance is compared to the likelihood at a variance of zero, and the setting with the larger likelihood is retained.

rescale_priors

If prior probabiltities are supplied, rescale them from 0-1 (i.e. rescaled_priors = priors / sum(priors)).

plot_track_fit

Record each iteration and make a GIF of the fine-mapping algorithm learning the causal variants. WARNING!: Making this plot can take a long time if there's many iterations.

verbose

If verbose = TRUE, the algorithm's progress, and a summary of the optimization settings, are printed to the console.

Details

Notes on convergence: susieR will often give the warning: IBSS algorithm did not converge in 100 iterations!. This means the results might not necessarily be reliable. There's several things you can try to avoid this:

Make sure susieR is up-to-date: devtools::install_github("stephenslab/susieR@0.9.0")

Increase max_causal (e.g. 5 => 10).

Increase max_iter (e.g. 100 => 1000), though this will take longer.

Decrease the locus window size, which will also speed up the algorithm but potentially miss causal variants far from the lead SNP.

Changing estimate_prior_method does not seem to affect covergence warnings.

Notes on variance: GitHub Issue If estimate_residual_variance=TRUE without providing var_y and L>1, susieR will throw error: Estimating residual variance failed: the estimated value is negative Running susieR with var_y = var(b) provides exactly the same results.

Examples

data("BST1"); data("LD_matrix");
# LD_matrix <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/BST1/plink/UKB_LD.RDS")
finemap_DT <- SUSIE(subset_DT=BST1, LD_matrix=LD_matrix, estimate_residual_variance=T)
#> [1] "+ Preparing sample_size (N) column"
#> [1] "++ Computing effective sample size."
#> [1] "+ SUSIE:: sample_size= 216621"
#> [1] "+ SUSIE:: max_causal = 5"
#> [1] "+ Subsetting LD matrix and finemap_dat to common SNPs..."
#> [1] "+ LD:: Removing unnamed rows/cols"
#> [1] "+ LD:: Replacing NAs with 0"
#> [1] "+ LD_matrix = 95 SNPs."
#> [1] "+ finemap_dat = 6216 SNPs."
#> [1] "+ 95 SNPs in common."
#> [1] "+ SUSIE:: Using `susie_suff_stat()` from susieR v0.11.92"
#> [1] "+ SUSIE:: Extracting Credible Sets..."