Merge fine-mapping results from all loci

Gather fine-mapping results from echolocatoR across all loci and merge into a single data.frame.

merge_finemapping_results(
  dataset = "./Data/GWAS",
  minimum_support = 1,
  include_leadSNPs = T,
  LD_reference = NULL,
  save_path = F,
  from_storage = T,
  haploreg_annotation = F,
  regulomeDB_annotation = F,
  biomart_annotation = F,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  exclude_methods = NULL,
  top_CS_only = F,
  verbose = T,
  nThread = 4
)

Arguments

dataset

Path to the folder you want to recursively search for results files within (e.g. "Data/GWAS/Nalls23andMe_2019"). Set this to a path that includes multiple subfolders if you want to gather results from multiple studies at once (e.g. "Data/GWAS").

minimum_support

Filter SNPs by the minimum number of fine-mapping tools that contained the SNP in their Credible Set.

include_leadSNPs

Include lead GWAS/QTL SNPs per locus (regardless of other filtering criterion).

from_storage

Search for stored results files.

haploreg_annotation

Annotate SNPs with HaploReg (using HaploR).

regulomeDB_annotation

Annotate SNPs with regulaomeDB (using HaploR).

biomart_annotation

Annotate SNPs with biomart.

PP_threshold

Mean posterior probability threshold to include SNPs in mean PP Credible Set (averaged across all fine-mapping tools).

exclude_methods

Exclude certain fine-mapping methods when estimating mean.CS and Consensus_SNP.

verbose

Print messages.

xlsx_path

Save merged data.frame as excel file.

consensus_thresh

The minimum number of tools that have the SNPs in their Credible Set to classify it as a Consensus_SNP.

Examples

dataset_dir <- "~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"
# UCS and lead SNPs: No annotation
merged_dat <- merge_finemapping_results(dataset=dataset_dir, minimum_support=1, include_leadSNPs=T)
#> [1] "+ Gathering all fine-mapping results from storage..."
#> Error in merge_finemapping_results(dataset = dataset_dir, minimum_support = 1,     include_leadSNPs = T): + No multi-finemap files found.

# UCS and lead SNPs: With annotations
merged_dat <- merge_finemapping_results(dataset=dataset_dir, minimum_support=1, include_leadSNPs=T, haploreg_annotation=T, biomart_annotation=T)
#> [1] "+ Gathering all fine-mapping results from storage..."
#> Error in merge_finemapping_results(dataset = dataset_dir, minimum_support = 1,     include_leadSNPs = T, haploreg_annotation = T, biomart_annotation = T): + No multi-finemap files found.